Moving RestAPIStoreUtils from rest.api.store -> rest.api.util package
